+++ src/test/scala/mesosphere/marathon/DebugConfTest.scala
@@ -5,21 +5,13 @@ import mesosphere.marathon.test.{ MarathonSpec, MarathonTestHelper }
 class DebugConfTest extends MarathonSpec {
   test("tracing is disabled by default") {
     val conf = MarathonTestHelper.defaultConfig()
-    assert(!conf.debugTracing())
+    assert(!conf.enableDebugTracing)
   }
 
   test("tracing can be enabled") {
     val conf = MarathonTestHelper.makeConfig(
       "--master", "127.0.0.1:5050",
       "--tracing"
-    )
-    assert(conf.enableDebugTracing)
-  }
-
-  test("tracing can be enabled using the deprecated ---enable_tracing") {
-    val conf = MarathonTestHelper.makeConfig(
-      "--master", "127.0.0.1:5050",
-      "--enable_tracing"
     )
     assert(conf.enableDebugTracing)
   }
@@ -34,7 +26,7 @@ class DebugConfTest extends MarathonSpec {
 
   test("metrics are enabled by default") {
     val conf = MarathonTestHelper.defaultConfig()
-    assert(conf.metrics())
+    assert(conf.enableMetrics)
   }
 
   test("metrics can be enabled") {
@@ -42,15 +34,7 @@ class DebugConfTest extends MarathonSpec {
       "--master", "127.0.0.1:5050",
       "--metrics"
     )
-    assert(conf.metrics())
-  }
-
-  test("the deprecated --enable_metrics is accepted") {
-    val conf = MarathonTestHelper.makeConfig(
-      "--master", "127.0.0.1:5050",
-      "--enable_metrics"
-    )
-    assert(conf.metrics())
+    assert(conf.enableMetrics)
   }
 
   test("metrics can be disabled") {
@@ -58,6 +42,6 @@ class DebugConfTest extends MarathonSpec {
       "--master", "127.0.0.1:5050",
       "--disable_metrics"
     )
-    assert(!conf.metrics())
+    assert(!conf.enableMetrics)
   }
 }